+++ CHANGELOG.md
@@ -105,6 +105,8 @@ ImportedLFP().drop()
     - Disable make transactionsfor `CuratedSpikeSorting` #1288
     - Refactor `SpikeSortingOutput.get_restricted_merge_ids` #1304
     - Add burst merge curation #1209
+    - Reconcile spikeinterface value for `channel_id` when `channel_name` column
+      present in nwb file electrodes table #1310
 - Behavior
     - Implement pipeline for keypoint-moseq extraction of behavior syllables #1056
 - LFP